Fixed up several issues.
